Refactored library code to include error checking.
Previously, had to copy errcheck.go into each module.

Now, all error checking functions (most importantly lib.Check) in lib.
These are exported as library functions.
