+++ gitcredhandler_test.go
@@ -8,7 +8,7 @@ import (
 
 func TestAddGitCredsHandler(t *testing.T) {
 	err := os.Setenv("CredentialsTest", "true")
-	check(err)
+	lib.Check(err)
 	// Git credentials shouldn't exist
 	if gitCredsExist() {
 		t.Error("Git credentials exist on entry into test (bad state)")
